+++ std/_util/has_own_property.ts
